the python edje bindings already encode to utf-8
The old code breaks on systems where the python default encoding is not utf-8
(check the default encoding with sys.getdefaultencoding()). Debian and Gentoo
have acsii as the default codec. So we need do explicitly encode and decode,
but at the right place (which is in the python bindings for edje in this case).

In OE we have a patch for python-edje which handles this, i'm trying to get
it upstream.