Give visual feedback when suspending
This will, in turn, highlight four bars on the screen: “keep pressed”,
“let go”, “suspending” and “waking up”. They are highlighted when the
user pushes down POWER, when he has pushed it for one second, when he
releases it in time and when the phone wakes up.

This also has the effect that keeping it pressed for more than three
seconds it will _not_ suspend, as to not interfere with a shutdown or
restart triggered by the Altr-Ctrl-Del sent this way.