Remove timer that automatically sends unsent messages
This was built into zhone at a time where sending SMS was flaky at best
and so resending messages that failed to send was a necessary evil.
Now this is not the case any more and this "feature" does more evil than
it helps. If anything we need a watchdog that keeps trying to send that
particular message until it has been sent successfully.