Avoid unnecessary deselection of the current SMS message
It is quite common for me to want to open an SMS before deleting or
replying to it.  Prior to this change, this requires selecting the
SMS, clicking the "open" button to see it, closing the message window,
then selecting the same SMS again, and so on.  This change removes the
need to select the same SMS again, by not resetting the selection when
the message window is closed.

* zhone (pyphone_sms::onShow): Don't call self.updateList().
  updateList() is already called in almost all the places that it's
  needed, i.e. cbMessagebookReply, and when an SMS is deleted.  There
  is one exception...

  (pyphone_contacts::cbPhonebookReply): Call the SMS page's
  updateList(), so as to remap phone numbers according to the newly
  read phonebook.